Move non TEE result encryption to post compute stage
